extract ItThinClientAbstractTest
